Bug #10384.  Fixed a problem that pops up because I fixed some exception handling a while back.
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@7910 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 811a112990301a74fa32d2c382c097478cf929ad

+++ tv/portable/iconcache.py
@@ -144,7 +144,8 @@ class IconCache:
             self.remove_file(self.filename)
 
     def reset(self):
-        self.remove_file(self.filename)
+        if self.filename:
+            self.remove_file(self.filename)
         self.filename = None
         self.resized_filenames = {}
         self.url = None
@@ -249,7 +250,7 @@ class IconCache:
 
             try:
                 fileutil.rename(tmp_filename, self.filename)
-            except:
+            except OSError:
                 self.filename = None
                 needsSave = True
